requirements.txt
requests~=2.25.1
Flask~=1.1.2
psycopg2-binary~=2.8.6
beautifulsoup4==4.9.1
gunicorn==20.1.0
requests==2.24.0
requests-oauthlib==1.3.0
